Raspberry Pi Web server
Set up the raspberry Pi webserver, removing the PIR aspect for now as it
needs testing further. This currently sets up a local server accessible
via devices on the network. Parameters 'on' or 'off' can be passed
through via a url, which enables and disables an LED.
The next step is to integrate the PIR sensor back in, send notifcations
once motion has been detected, and make the server accessible via the
internet (not just local network) by port forwarding.