Shared preferences for device links
Added the ability to link the app to a raspberry pi motion detector
device, saving the connection info (ip & port) in shared preferences.
This reduces how many times a user has to input the information. Also
this adds room for expansion in the future so multiple devices could be
added etc.